PR feedback on complex and file headers.
